refactor: split eslint checks
