+++ eslint.config.mjs
@@ -10,205 +10,11 @@ import { reactRefresh } from 'eslint-plugin-react-refresh'
 import storybook from 'eslint-plugin-storybook'
 import unusedImports from 'eslint-plugin-unused-imports'
 import tseslint from 'typescript-eslint'
+import { useMaybePatternRule } from './.eslint/use-maybe-pattern.rule.mjs'
+import { noDoubleNegativeRule } from './.eslint/no-double-negative.rule.mjs'
+import { noRedundantTernaryRule } from './.eslint/no-redundant-ternary.rule.mjs'
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
-
-/**
- * Custom rule that guards for usages of the `maybe` helper.
- * @type {eslint.Rule.Module}
- **/
-const noMaybePatternRule = {
-  meta: {
-    type: 'suggestion',
-    docs: {
-      description: 'Disallow null-check patterns that should use the `maybe` helper from @primitives/objects.utils',
-    },
-    messages: {
-      ternary: 'Use `maybe(value, fn)` instead of a ternary with a null check. See @primitives/objects.utils.',
-      guard: 'Use `return maybe(value, fn)` instead of an early-return null guard. See @primitives/objects.utils.',
-    },
-  },
-  create(context) {
-    /** Returns 'null' | 'undefined' if the node is a nullish literal, or false otherwise */
-    const isNullish = node =>
-      (node?.type === 'Literal' && node.value === null) || (node?.type === 'Identifier' && node.name === 'undefined')
-
-    /**
-     * Check if all null checks in both sides of a logical expression satisfy the given predicate, and if so,
-     * return a combined array of all checks. Otherwise, return null.
-     **/
-    const isEveryNullCheck = ([leftChecks, rightChecks], predicate) =>
-      leftChecks?.every(predicate) && rightChecks?.every(predicate) ? [...leftChecks, ...rightChecks] : null
-
-    /**
-     * Recursively extracts null checks from a condition expression.
-     * Returns an array of { checkedValue, isNegated } objects, or null if the
-     * condition is not purely composed of null checks.
-     *
-     * Supported patterns:
-     *   x == null, x === null, x != null, x !== null  (also with undefined)
-     *   a == null || b == null    (all affirmative)
-     *   a != null && b != null    (all negated)
-     */
-    const getNullChecks = ({ left, operator, right, type }) =>
-      type === 'BinaryExpression'
-        ? ['==', '===', '!=', '!=='].includes(operator)
-          ? isNullish(right)
-            ? [{ checkedValue: left, isNegated: operator.includes('!') }]
-            : isNullish(left)
-              ? [{ checkedValue: right, isNegated: operator.includes('!') }]
-              : null
-          : null
-        : type === 'LogicalExpression'
-          ? operator === '||'
-            ? // All checks in an || chain must be affirmative (== null)
-              isEveryNullCheck([getNullChecks(left), getNullChecks(right)], c => !c.isNegated)
-            : operator === '&&'
-              ? // All checks in an && chain must be negated (!= null)
-                isEveryNullCheck([getNullChecks(left), getNullChecks(right)], c => c.isNegated)?.map(c => ({
-                  ...c,
-                  isNegated: false, // unify != null && and == null || since they are equivalent
-                }))
-              : null
-          : null
-
-    const isThenNullish = ({ argument, body, type }) =>
-      type === 'ReturnStatement'
-        ? isNullish(argument)
-        : type === 'BlockStatement' &&
-          body.length === 1 &&
-          body[0].type === 'ReturnStatement' &&
-          isNullish(body[0].argument)
-
-    return {
-      /** Detect ternary patterns: x == null ? nullish : expr  or  x != null ? expr : nullish */
-      ConditionalExpression: node =>
-        // Exactly one branch must be nullish && The entire condition must be composed of null checks
-        isNullish(node.consequent) !== isNullish(node.alternate) &&
-        getNullChecks(node.test) &&
-        context.report({ node, messageId: 'ternary' }),
-      /** Detect guard patterns: if (x == null) return nullish; return expr */
-      IfStatement(node) {
-        const { test, parent, consequent } = node
-        const nullChecks = getNullChecks(test)
-        if (
-          // Condition must be a single affirmative null check (not negated)
-          nullChecks?.length === 1 &&
-          !nullChecks[0].isNegated &&
-          // The then-block must be a return with a nullish value
-          isThenNullish(consequent) &&
-          // the next sibling must be a return statement
-          (parent.type === 'BlockStatement' || parent.type === 'Program') &&
-          parent.body[parent.body.indexOf(node) + 1]?.type === 'ReturnStatement'
-        )
-          context.report({
-            node,
-            messageId: 'guard',
-          })
-      },
-    }
-  },
-}
-
-/** Forbids negated ternary conditions like `!x ? a : b`, which should be `x ? b : a`. */
-const noDoubleNegativeRule = {
-  meta: {
-    type: 'suggestion',
-    docs: { description: 'Disallow negated conditions in ternary expressions' },
-    messages: { negated: 'Negated ternary conditions are hard to read. Swap the branches and remove the `!`.' },
-  },
-  create: context => ({
-    ConditionalExpression(node) {
-      const { test } = node
-      if (test.type === 'UnaryExpression' && test.operator === '!') context.report({ node: test, messageId: 'negated' })
-    },
-  }),
-}
-
-/** Forbids redundant ternary patterns like `x ? x : y`, which should be `x || y`. */
-const noRedundantTernaryRule = {
-  meta: {
-    type: 'suggestion',
-    docs: { description: 'Disallow redundant ternary expressions that can be simplified to logical OR' },
-    messages: { redundant: 'Prefer `{{test}} || {{alternate}}` instead of `{{test}} ? {{test}} : {{alternate}}`.' },
-  },
-  create: context => {
-    /**
-     * Check if two AST nodes are structurally equivalent.
-     * Returns true if both nodes represent the same source code structure.
-     */
-    const areNodesEqual = (node1, node2) => {
-      if (!node1 || !node2) return node1 === node2
-      if (node1.type !== node2.type) return false
-      switch (node1.type) {
-        case 'Identifier':
-          return node1.name === node2.name
-        case 'Literal':
-          return node1.value === node2.value
-        case 'MemberExpression':
-          return (
-            areNodesEqual(node1.object, node2.object) &&
-            areNodesEqual(node1.property, node2.property) &&
-            node1.computed === node2.computed
-          )
-        case 'ChainExpression':
-          return areNodesEqual(node1.expression, node2.expression)
-        case 'CallExpression':
-          return (
-            areNodesEqual(node1.callee, node2.callee) &&
-            node1.arguments.length === node2.arguments.length &&
-            node1.arguments.every((arg, i) => areNodesEqual(arg, node2.arguments[i]))
-          )
-        case 'ArrayExpression':
-          return (
-            node1.elements.length === node2.elements.length &&
-            node1.elements.every((el, i) => areNodesEqual(el, node2.elements[i]))
-          )
-        case 'ObjectExpression':
-          return (
-            node1.properties.length === node2.properties.length &&
-            node1.properties.every((prop, i) => areNodesEqual(prop, node2.properties[i]))
-          )
-        case 'Property':
-          return areNodesEqual(node1.key, node2.key) && areNodesEqual(node1.value, node2.value)
-        case 'UnaryExpression':
-          return node1.operator === node2.operator && areNodesEqual(node1.argument, node2.argument)
-        case 'BinaryExpression':
-        case 'LogicalExpression':
-          return (
-            node1.operator === node2.operator &&
-            areNodesEqual(node1.left, node2.left) &&
-            areNodesEqual(node1.right, node2.right)
-          )
-        case 'ConditionalExpression':
-          return (
-            areNodesEqual(node1.test, node2.test) &&
-            areNodesEqual(node1.consequent, node2.consequent) &&
-            areNodesEqual(node1.alternate, node2.alternate)
-          )
-        default:
-          return false
-      }
-    }
-
-    return {
-      ConditionalExpression(node) {
-        const { test, consequent, alternate } = node
-        // Only report if test and consequent are identical
-        if (areNodesEqual(test, consequent)) {
-          context.report({
-            node,
-            messageId: 'redundant',
-            data: {
-              test: context.sourceCode.getText(test),
-              alternate: context.sourceCode.getText(alternate),
-            },
-          })
-        }
-      },
-    }
-  },
-}
 
 const config = [
   // Global ignores — must be its own object with no `files` key
@@ -244,7 +50,7 @@ const config = [
       'unused-imports': unusedImports,
       local: {
         rules: {
-          'use-maybe-pattern': noMaybePatternRule,
+          'use-maybe-pattern': useMaybePatternRule,
           'no-double-negative': noDoubleNegativeRule,
           'no-redundant-ternary': noRedundantTernaryRule,
         },
