+++ packages/prices-api/src/pools/schema.ts
@@ -328,11 +328,11 @@ export const getPoolMetadataResponse = z
     has_donations: z.boolean(),
   })
   .transform(camelizeKeys)
-  .transform(({ deploymentDate, gauges, assetTypes, oracles, ...data }) => ({
+  .transform(({ deploymentDate, gauges, assetTypes, oracles = null, ...data }) => ({
     ...data,
     gauges: [...gauges],
     assetTypes: assetTypes ? [...assetTypes] : null,
-    oracles: oracles?.map(item => (item ? item : null)) ?? null,
+    oracles,
     deploymentDate: deploymentDate ?? null,
   }))
